requirements.txt
fastapi==0.115.0
uvicorn==0.32.0
motor==3.6.0
passlib==1.7.4
bcrypt==4.2.0
python-jose==3.3.0
python-dotenv==1.0.1

# YANGI QO'SHILGAN - XAVFSIZLIK
slowapi              # Rate limiting
redis               # Redis cache
pydantic[email]     # Input validation
python-magic        # File type detection
Pillow              # Image processing